+++ Public_API/Interfaces/ITrades.cs
@@ -6,8 +6,8 @@
         int TradeId { get;}
         string Type { get; }
         decimal Price { get; }
-        decimal Quantity { get;}
-        decimal Amount { get; }
+        double Quantity { get;}
+        double Amount { get; }
         double Date { get; }
     }
